feat(sidebar): change color change arch
