--- src/components/logo/Logo.module.scss
@@ -1,22 +1,0 @@
-.logo {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    height: 40px;
-    margin-bottom: 30px;
-    span {
-        color: #f6f7fb;
-        font-weight: 800;
-        display: inline-block;
-        padding: 0 3px;
-    }
-    .icon {
-        font-size: 45px;
-        color: rgb(57, 209, 166);
-    }
-    .active {
-        font-size: 30px;
-        color: rgb(57, 209, 166);
-    }
-}
