+++ src/components/tabs/Tabs.tsx
@@ -1,12 +1,14 @@
 import { Suspense, useState } from 'react';
+
+import { Tab } from '@/widgets/sidebar/model/types';
+
 import Loader from '../ui/loader/Loader';
-import { Tab } from '@/types/navigation';
 import styles from './Tabs.module.scss';
 
 interface TabsProps {
   tabs: Tab[];
 }
-const Tabs = ({ tabs }:TabsProps) => {
+const Tabs = ({ tabs }: TabsProps) => {
   const [activeTab, setActiveTab] = useState<number>(tabs[0].id);
   return (
     <>
@@ -14,22 +16,20 @@ const Tabs = ({ tabs }:TabsProps) => {
         {tabs.map((tab) => (
           <button
             key={tab.id}
-            className={`${styles["tab-button"]} ${activeTab === tab.id 
-                ? styles.active
-                : ""
-            }`}
-            onClick={() => setActiveTab(tab.id)}>
+            className={`${styles['tab-button']} ${activeTab === tab.id ? styles.active : ''}`}
+            onClick={() => setActiveTab(tab.id)}
+          >
             {tab.label}
           </button>
         ))}
       </div>
-      <div className={styles["tab-content"]}>
+      <div className={styles['tab-content']}>
         {tabs.map((tab) =>
-          tab.id === activeTab 
-          ? <Suspense fallback={<Loader color="blue" size="sm" />} key={tab.id}>
-                <tab.component />
+          tab.id === activeTab ? (
+            <Suspense fallback={<Loader color="blue" size="sm" />} key={tab.id}>
+              <tab.component />
             </Suspense>
-          : null
+          ) : null
         )}
       </div>
     </>
