+++ src/routes/Layout.tsx
@@ -1,13 +1,16 @@
 import { useState } from 'react';
+
+import { ErrorBoundary } from 'react-error-boundary';
 import { Outlet, useLocation } from 'react-router-dom';
+
+import LocalFallback from '@/components/ui/error/LocalFallback';
+
+import Breadcrumbs from '../components/breadcrumbs/Breadcrumbs';
 import Footer from '../components/footer/Footer';
 import Header from '../components/header/Header';
-import Sidebar from '../components/sidebar/Sidebar';
-import { ErrorBoundary } from 'react-error-boundary';
-import AdminMenu from '../components/navigates/admin/AdminMenu';
 import { GlobalModal } from '../components/modal/GlobalModal';
-import LocalFallback from '@/components/ui/error/LocalFallback';
-import Breadcrumbs from '../components/breadcrumbs/Breadcrumbs';
+import AdminMenu from '../components/navigates/admin/AdminMenu';
+import Sidebar from '../widgets/sidebar/ui/Sidebar/Sidebar';
 import styles from './Layout.module.scss';
 
 const Layout = () => {
@@ -17,11 +20,11 @@ const Layout = () => {
   return (
     <div className={styles.main}>
       <Sidebar open={isActive} />
-      <div className={isActive ? styles.active : styles["not-active"]}>
+      <div className={isActive ? styles.active : styles['not-active']}>
         <Header isActive={isActive} setIsActive={setIsActive} />
         <Breadcrumbs />
         <main className={styles.outlet}>
-          {location.pathname.startsWith("/admin") && <AdminMenu />}
+          {location.pathname.startsWith('/admin') && <AdminMenu />}
           <div className={styles.content}>
             <ErrorBoundary FallbackComponent={LocalFallback}>
               <Outlet />
