+++ src/widgets/sidebar/ui/Logo/Logo.module.scss
@@ -1,0 +1,22 @@
+.logo {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 40px;
+  margin-bottom: 30px;
+  span {
+    color: #f6f7fb;
+    font-weight: 800;
+    display: inline-block;
+    padding: 0 3px;
+  }
+  .icon {
+    font-size: 45px;
+    color: var(--orange-500);
+  }
+  .active {
+    font-size: 30px;
+    color: var(--gray-400);
+  }
+}
