+++ src/widgets/sidebar/ui/Logo/Logo.tsx
@@ -1,10 +1,12 @@
-import { MdBarcodeReader } from "react-icons/md";
-import { IoBarcodeOutline } from "react-icons/io5";
-import { ELEMENTS_LABELS } from "@/utils/constants/ui/elements";
+import { IoBarcodeOutline } from 'react-icons/io5';
+import { MdBarcodeReader } from 'react-icons/md';
+
+import { ELEMENTS_LABELS } from '@/utils/constants/ui/elements';
+
 import styles from './Logo.module.scss';
 
 interface LogoProps {
-    open: boolean;
+  open: boolean;
 }
 
 const Logo = ({ open }: LogoProps) => {
